+++ build_daemon/lib/data/build_target.dart
@@ -23,5 +23,21 @@ abstract class DefaultBuildTarget
   BuiltSet<RegExp> get blackListPatterns;
 
   @nullable
+  OutputLocation get outputLocation;
+}
+
+abstract class OutputLocation
+    implements Built<OutputLocation, OutputLocationBuilder> {
+  static Serializer<OutputLocation> get serializer =>
+      _$outputLocationSerializer;
+
+  factory OutputLocation([updates(OutputLocationBuilder b)]) = _$OutputLocation;
+
+  OutputLocation._();
+
   String get output;
+
+  bool get useSymlinks;
+
+  bool get hoist;
 }