+++ build_daemon/lib/data/serializers.g.dart
@@ -15,6 +15,7 @@ Serializers _$serializers = (new Serializers().toBuilder()
       ..add(DefaultBuildTarget.serializer)
       ..add(OutputLocation.serializer)
       ..add(ServerLog.serializer)
+      ..add(ShutdownNotification.serializer)
       ..addBuilderFactory(
           const FullType(BuiltList, const [const FullType(BuildResult)]),
           () => new ListBuilder<BuildResult>())